+++ include/ui/LibraryUI.h
@@ -11,5 +11,5 @@ public:
 
     virtual void displayMenu() = 0;
 
-    virtual LibraryItem gatherMissingInfoInteractiveUI(long int id, std::string name, std::string author, std::string description, std::string kindIsOrString) = 0;
+    virtual LibraryItem gatherMissingInfoInteractiveUI(long long id, std::string name, std::string author, std::string description, std::string kindIsOrString) = 0;
 };
